mindful-living-backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-filter==23.5
Pillow==10.2.0
django-storages==1.14.2
django-rest-auth==0.9.5
django-allauth==0.61.1
channels==4.0.0
channels-redis==4.2.0 